LayoutTests/imported/w3c/web-platform-tests/html/syntax/speculative-parsing/generated/document-write/svg-image-xlinkhref.tentative.sub-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_not_equals: speculative case did not fetch got disallowed value ""

    <!-- speculative case in document.write -->
    <svg><image xlink:href="/html/syntax/speculative-parsing/resources/stash.py?action=put&amp;uuid=c99f657b-adb9-491d-ad95-2778e5cefc32&amp;encodingcheck=&Gbreve;"></image></svg>


FAIL Speculative parsing, document.write(): svg-image-xlinkhref Unhandled rejection: assert_not_equals: speculative case did not fetch got disallowed value ""
